Initial code created
I began the documentation and created some code to represent the neural network. This included initialisation of the object and generating the starting weights and biases. It also includes prediction. I will next session change the structure of the code to reduce the responsibilities of the main class by creating a class to represent a layer transformation. The model will then have reduced responsibilities as it will be composed of these layers.